% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{balance_dataset}
\alias{balance_dataset}
\title{Balance training dataset}
\usage{
balance_dataset(mat, tag)
}
\arguments{
\item{mat}{count matrix of dimension m x n,
corresponding to m cells and n features}

\item{tag}{named list of training tags/labels (yes/no)
corresponding to a specific cell type, name and length of
list must be coherent with cells in mat}
}
\value{
a list of balanced count matrix
and corresponding tags of balanced count matrix
}
\description{
Balance training dataset
}
