% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{cell_type<-,SingleCellClassR-method}
\alias{cell_type<-,SingleCellClassR-method}
\title{Setter for cell_type
Change cell type for a classifier}
\usage{
\S4method{cell_type}{SingleCellClassR}(classifier) <- value
}
\arguments{
\item{classifier}{SingleCellClassR object. 
The object is returned from the train_classifier function.}

\item{value}{the new cell type}
}
\value{
SingleCellClassR object with the new cell type.
}
\description{
Setter for cell_type
Change cell type for a classifier
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_B, cell_type = "B cells")
cell_type(clf_b) <- "B cell"
}
