% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{cell_type<-}
\alias{cell_type<-}
\title{Setter for cell_type.
Change cell type of a classifier}
\usage{
cell_type(classifier) <- value
}
\arguments{
\item{classifier}{the classifier whose cell type will be changed}

\item{value}{the new cell type}
}
\value{
the classifier with the new cell type
}
\description{
Setter for cell_type.
Change cell type of a classifier
}
