% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{check_parent_child_coherence}
\alias{check_parent_child_coherence}
\alias{check_parent_child_coherence,Seurat-method}
\alias{check_parent_child_coherence,SingleCellExperiment-method}
\title{Check label coherence in parent and child cell type}
\usage{
check_parent_child_coherence(
  obj,
  pos_parent,
  parent_cell,
  cell_type,
  target_cell_type,
  ...
)

\S4method{check_parent_child_coherence}{Seurat}(
  obj,
  pos_parent,
  parent_cell,
  cell_type,
  target_cell_type,
  tag_slot
)

\S4method{check_parent_child_coherence}{SingleCellExperiment}(
  obj,
  pos_parent,
  parent_cell,
  cell_type,
  target_cell_type,
  tag_slot
)
}
\arguments{
\item{obj}{object}

\item{pos_parent}{a vector indicating parent clf prediction}

\item{parent_cell}{name of parent cell type}

\item{cell_type}{name of child cell type}

\item{target_cell_type}{alternative cell types (in case of testing clf)}

\item{...}{arguments passed to other methods}

\item{tag_slot}{tag slot in \code{\link{SingleCellExperiment}} object 
indicating cell type}
}
\value{
list of adjusted object and adjusted tag slot
}
\description{
Check label coherence in parent and 
child cell type in a \code{\link{Seurat}} object.

Check label coherence in parent and child cell type 
in a \code{\link{SingleCellExperiment}} object
}
