% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{checkCellTypeValidity}
\alias{checkCellTypeValidity}
\title{Check validity of classifier cell type.}
\usage{
checkCellTypeValidity(cell_type)
}
\arguments{
\item{cell_type}{Classifier cell type to check.}
}
\value{
TRUE if the cell type is valid or the reason why it is not.
}
\description{
Check validity of classifier cell type.
}
