% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{checkClassifierValidity}
\alias{checkClassifierValidity}
\title{Check validity of classifier}
\usage{
checkClassifierValidity(clf)
}
\arguments{
\item{clf}{Classifier to check.}
}
\value{
TRUE if the classifier is valid or the reason why it is not.
}
\description{
Check validity of classifier
}
