% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{checkFeaturesValidity}
\alias{checkFeaturesValidity}
\title{Check validity of classifier features.}
\usage{
checkFeaturesValidity(features)
}
\arguments{
\item{features}{Classifier features to check.}
}
\value{
TRUE if the features is valid or the reason why it is not.
}
\description{
Check validity of classifier features.
}
