% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{checkObjectValidity}
\alias{checkObjectValidity}
\title{Check if a scTypeR object is valid}
\usage{
checkObjectValidity(object)
}
\arguments{
\item{object}{The request classifier to check.}
}
\value{
TRUE if the classifier is valid or the reason why it is not
}
\description{
Check if a scTypeR object is valid
}
