% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{checkParentValidity}
\alias{checkParentValidity}
\title{Check validity of classifier parent.}
\usage{
checkParentValidity(parent)
}
\arguments{
\item{parent}{Classifier parent to check.}
}
\value{
TRUE if the parent is valid or the reason why it is not.
}
\description{
Check validity of classifier parent.
}
