% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{checkPThresValidity}
\alias{checkPThresValidity}
\title{Check validity of classifier probability threshold.}
\usage{
checkPThresValidity(p_thres)
}
\arguments{
\item{p_thres}{Classifier probability threshold to check.}
}
\value{
TRUE if the p_thres is valid or the reason why it is not.
}
\description{
Check validity of classifier probability threshold.
}
