% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{clf<-,SingleCellClassR-method}
\alias{clf<-,SingleCellClassR-method}
\title{Setter for clf.
Change of clf will also lead to change of features.}
\usage{
\S4method{clf}{SingleCellClassR}(classifier) <- value
}
\arguments{
\item{classifier}{SingleCellClassR object. 
The object is returned from the train_classifier function.}

\item{value}{the new classifier}
}
\value{
SingleCellClassR object with the new trained classifier.
}
\description{
Setter for clf.
Change of clf will also lead to change of features.
}
