% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{clf<-}
\alias{clf<-}
\title{Setter for clf.
Change of clf will also lead to change of features.}
\usage{
clf(classifier) <- value
}
\arguments{
\item{classifier}{scTypeR object. 
The object is returned from the train_classifier function.}

\item{value}{the new classifier}
}
\value{
scTypeR object with the new trained classifier.
}
\description{
Setter for clf.
Change of clf will also lead to change of features.
}
