% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{clf,SingleCellClassR-method}
\alias{clf,SingleCellClassR-method}
\title{clf}
\usage{
\S4method{clf}{SingleCellClassR}(classifier)
}
\arguments{
\item{classifier}{\code{\link{SingleCellClassR}} object}
}
\value{
Classifier is the object returned by caret SVM learning process.
More information about the caret package: https://topepo.github.io/caret/
}
\description{
Returns the classifier of the \code{\link{SingleCellClassR}} object
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_B, cell_type = "B cells")
clf(clf_b)

}
