% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{construct_tag_vect}
\alias{construct_tag_vect}
\alias{construct_tag_vect,Seurat-method}
\alias{construct_tag_vect,SingleCellExperiment-method}
\title{Construct tag vector}
\usage{
construct_tag_vect(obj, cell_type, ...)

\S4method{construct_tag_vect}{Seurat}(obj, cell_type, tag_slot)

\S4method{construct_tag_vect}{SingleCellExperiment}(obj, cell_type, tag_slot)
}
\arguments{
\item{obj}{object}

\item{cell_type}{name of cell type}

\item{...}{arguments passed to other methods}

\item{tag_slot}{tag slot in \code{\link{SingleCellExperiment}} object 
indicating cell type}
}
\value{
a binary vector for cell tag
}
\description{
Construct a uniform tag vector for all forms of labels 
in a \code{\link{Seurat}} object

Construct a uniform tag vector for all forms of labels
in a \code{\link{SingleCellExperiment}} object
}
