% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_models}
\alias{default_models}
\title{Pretrained classifiers for human cells}
\format{
a list of \code{\link{scAnnotatR}} objects
}
\usage{
default_models
}
\description{
Pretrained classifier obtained by training and testing on the 
Sade-Feldman melanoma dataset, the Jerby-Arnon melanoma dataset, the Haniffa
Skin Cell Atlas and the Haniffa Covid-19 Cell Atlas.
}
\author{
Vy Nguyen, June 2021
}
\keyword{datasets}
