% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_models_m}
\alias{default_models_m}
\title{Pretrained classifiers for murine cells}
\format{
a list of \code{\link{SingleCellClassR}} objects
}
\usage{
default_models_n
}
\description{
Pretrained classifier obtained by training and testing on 
The Tabula Muris Consortium.
}
\author{
Vy Nguyen, November 2020
}
\keyword{datasets}
