% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{features,scTypeR-method}
\alias{features,scTypeR-method}
\title{features}
\usage{
\S4method{features}{scTypeR}(classifier)
}
\arguments{
\item{classifier}{scTypeR object}
}
\value{
Applied features of object
}
\description{
Returns the set of features for the given classifier.
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_B, cell_type = "B cells")
features(clf_b)

}
