% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{features<-,scTypeR-method}
\alias{features<-,scTypeR-method}
\title{Setter for features. Users are not allowed to change features.}
\usage{
\S4method{features}{scTypeR}(classifier) <- value
}
\arguments{
\item{classifier}{scTypeR object. 
The object is returned from the train_classifier function.}

\item{value}{the new classifier}
}
\value{
scTypeR object with the new features.
}
\description{
Setter for features. Users are not allowed to change features.
}
