% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{features<-,SingleCellClassR-method}
\alias{features<-,SingleCellClassR-method}
\title{Setter for features. Users are not allowed to change features.}
\usage{
\S4method{features}{SingleCellClassR}(classifier) <- value
}
\arguments{
\item{classifier}{SingleCellClassR object. 
The object is returned from the train_classifier function.}

\item{value}{the new classifier}
}
\value{
SingleCellClassR object with the new features.
}
\description{
Setter for features. Users are not allowed to change features.
}
