% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{highest_expressed_ctype}
\alias{highest_expressed_ctype}
\title{Get the cell type having the highest average expression}
\usage{
highest_expressed_ctype(types, cell_exp, classifiers)
}
\arguments{
\item{types}{list of possible cell types}

\item{cell_exp}{cell expression}

\item{classifiers}{classifiers}
}
\value{
cell type having the highest average expression
}
\description{
Get the cell type having the highest average expression
}
