% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{load_models}
\alias{load_models}
\title{Load classifiers from databases}
\usage{
load_models(path_to_models)
}
\arguments{
\item{path_to_models}{path to databases, or by default}
}
\value{
list of classifiers
}
\description{
Load classifiers from databases
}
