% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{make_prediction}
\alias{make_prediction}
\title{Make prediction}
\usage{
make_prediction(mat, classifier, pred_cells, ignore_ambiguous_result = TRUE)
}
\arguments{
\item{mat}{count matrix used for prediction}

\item{classifier}{classifier}

\item{pred_cells}{a whole prediction for all cells}

\item{ignore_ambiguous_result}{whether ignore ambigouous result}
}
\value{
prediction
}
\description{
Make prediction
}
