% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{p_thres<-}
\alias{p_thres<-}
\title{Setter for predicting probability threshold}
\usage{
p_thres(classifier) <- value
}
\arguments{
\item{classifier}{the classifier whose predicting probability threshold 
will be changed}

\item{value}{the new threshold}
}
\value{
classifier with the new threshold.
}
\description{
Setter for predicting probability threshold
}
