% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{parent,scTypeR-method}
\alias{parent,scTypeR-method}
\title{parent}
\usage{
\S4method{parent}{scTypeR}(classifier)
}
\arguments{
\item{classifier}{scTypeR object}
}
\value{
Parent model of object
}
\description{
Returns the parent of the cell type corresponding to the given classifier.
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_B, cell_type = "B cells")
parent(clf_b)

}
