% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{parent<-,scTypeR-method}
\alias{parent<-,scTypeR-method}
\title{Setter for parent}
\usage{
\S4method{parent}{scTypeR}(classifier) <- value
}
\arguments{
\item{classifier}{scTypeR object. 
The object is returned from the train_classifier function.}

\item{value}{the new parent}
}
\value{
scTypeR object with the new parent.
}
\description{
Setter for parent
}
