% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{parent<-}
\alias{parent<-}
\title{Setter for parent}
\usage{
parent(classifier) <- value
}
\arguments{
\item{classifier}{SingleCellClassR object. 
The object is returned from the train_classifier function.}

\item{value}{the new parent}
}
\value{
SingleCellClassR object with the new parent.
}
\description{
Setter for parent
}
