% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{parent}
\alias{parent}
\title{parent}
\usage{
parent(classifier)
}
\arguments{
\item{classifier}{scAnnotatR object}
}
\value{
Parent model of object
}
\description{
Returns the parent of the cell type corresponding to the given classifier.
}
\examples{
data("tirosh_mel80_example")
selected_marker_genes_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
classifier_b <- train_classifier(train_obj = tirosh_mel80_example,
assay = 'RNA', layer = 'counts', marker_genes = selected_marker_genes_B, 
cell_type = "B cells", tag_slot = 'active.ident')
parent(classifier_b)

}
