% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{process_parent_clf}
\alias{process_parent_clf}
\alias{process_parent_clf,Seurat-method}
\alias{process_parent_clf,SingleCellExperiment-method}
\title{Process parent clf}
\usage{
process_parent_clf(
  obj,
  parent_tag_slot,
  parent_cell_type,
  parent_clf,
  path_to_models,
  zscore = TRUE,
  ...
)

\S4method{process_parent_clf}{Seurat}(
  obj,
  parent_tag_slot,
  parent_cell_type,
  parent_clf,
  path_to_models,
  zscore = TRUE,
  seurat_assay,
  seurat_slot,
  ...
)

\S4method{process_parent_clf}{SingleCellExperiment}(
  obj,
  parent_tag_slot,
  parent_cell_type,
  parent_clf,
  path_to_models,
  zscore = TRUE,
  sce_assay,
  ...
)
}
\arguments{
\item{obj}{object}

\item{parent_tag_slot}{string, name of annotation tag slot in object 
indicating pre-assigned/predicted parent cell type}

\item{parent_cell_type}{name of parent cell type}

\item{parent_clf}{\code{\link{SingleCellClassR}} object corresponding 
to classification model for the parent cell type}

\item{path_to_models}{path to databases, or by default}

\item{zscore}{boolean indicating the transformation of gene expression 
in object to zscore or not}

\item{...}{arguments passed to other methods}

\item{seurat_assay}{name of assay to use in \code{\link{Seurat}} object}

\item{seurat_slot}{type of expression data to use in 
\code{\link{Seurat}} object}

\item{sce_assay}{name of assay to use 
in \code{\link{SingleCellExperiment}} object}
}
\value{
list of cells which are positive to parent clf
}
\description{
Process parent classifier in a \code{\link{Seurat}} object

Process parent classifier in 
a \code{\link{SingleCellExperiment}} object
}
