% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{select_features}
\alias{select_features}
\title{Perform features selection and handle missing features}
\usage{
select_features(mat, features)
}
\arguments{
\item{mat}{count matrix of dimension n x m 
corresponding to m cells and n features}

\item{features}{list of selected features}
}
\value{
filtered matrix
}
\description{
Perform features selection and handle missing features
}
