% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{SingleCellClassR}
\alias{SingleCellClassR}
\title{SingleCellClassR class.}
\usage{
SingleCellClassR(cell_type, clf, features, p_thres, parent)

SingleCellClassR(cell_type, clf, features, p_thres, parent)
}
\arguments{
\item{cell_type}{character. Name of the cell type.}

\item{clf}{list. Trained model returned by caret train function.}

\item{features}{vector/character containing features used for the training.}

\item{p_thres}{numeric. 
Probability threshold for the cell type to be assigned for a cell.}

\item{parent}{character. Parent cell type.}
}
\value{
A SingleCellClassR object.
}
\description{
This class is returned by the \code{\link{train_classifier}} function.
Generally, SingleCellClassR objects are never created directly.
}
\section{Slots}{

\describe{
\item{\code{cell_type}}{character. Name of the cell type.}

\item{\code{clf}}{list. Trained model returned by caret train function.}

\item{\code{features}}{vector/character containing features 
used for the training.}

\item{\code{p_thres}}{numeric. 
Probability threshold for the cell type to be assigned for a cell.}

\item{\code{parent}}{character. Parent cell type.}
}}

\examples{
# load small example dataset
data("tirosh_mel80_example")

# train a classifier, for ex: B cell
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
                          features = selected_features_B, cell_type = "B cells")

clf_b
}
