% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{test_performance}
\alias{test_performance}
\title{Test clf performance}
\usage{
test_performance(mat, classifier, tag)
}
\arguments{
\item{mat}{expression matrix}

\item{classifier}{classifier}

\item{tag}{tag of data}
}
\value{
clf performance
}
\description{
Test clf performance
}
