% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tirosh_mel80_example}
\alias{tirosh_mel80_example}
\title{A Seurat Object Sample}
\format{
a \code{\link{Seurat}} object
}
\source{
WEIZMANN INSTITUTE OF SCIENCE
}
\usage{
tirosh_mel80_example
}
\description{
An example Seurat object shipped with the package as an example
data. The expression data was originally from the dataset GSE72056, with
samples corresponding to patient CY80. The Seurat object was then adapted to
be used in \code{\link{scAnnotatR}}.
}
\author{
Itay Tirosh, 2016-04-05
}
\keyword{datasets}
