% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{transform_to_zscore}
\alias{transform_to_zscore}
\title{Transform whole matrix of counts to z-score}
\usage{
transform_to_zscore(mat)
}
\arguments{
\item{mat}{count matrix of dimension m x n 
corresponding to m cells and n features}
}
\value{
row wise center-scaled count matrix
}
\description{
Transform whole matrix of counts to z-score
}
