% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{verify_parent}
\alias{verify_parent}
\title{Verify parent prediction}
\usage{
verify_parent(mat, classifier, meta.data)
}
\arguments{
\item{mat}{expression matrix}

\item{classifier}{classifier}

\item{meta.data}{object meta data}
}
\value{
applicable matrix
}
\description{
Verify parent prediction
}
