% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{visualize_tree,data.frame-method}
\alias{visualize_tree,data.frame-method}
\title{View cell type tree}
\usage{
\S4method{visualize_tree}{data.frame}(tree)
}
\arguments{
\item{tree}{data frame, tree to be visualize}
}
\value{
visualization of the tree in console
}
\description{
View cell type tree
}
\examples{
t <- plant_tree()
visualize_tree(t)
}
