% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{caret_model}
\alias{caret_model}
\title{caret_model}
\usage{
caret_model(classifier)
}
\arguments{
\item{classifier}{\code{\link{scAnnotatR}} object}
}
\value{
Classifier is the object returned by caret SVM learning process.
More information about the caret package: https://topepo.github.io/caret/
}
\description{
Returns the caret model of the \code{\link{scAnnotatR}} object
}
\examples{
data("tirosh_mel80_example")
selected_marker_genes_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
classifier_b <- train_classifier(train_obj = tirosh_mel80_example, 
marker_genes = selected_marker_genes_B, cell_type = "B cells")
caret_model(classifier_b)
 
}
