% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{cell_type}
\alias{cell_type}
\alias{cell_type,scTypeR-method}
\title{cell_type}
\usage{
cell_type(classifier)

\S4method{cell_type}{scTypeR}(classifier)
}
\arguments{
\item{classifier}{\code{\link{scTypeR}} object}
}
\value{
cell type of object
}
\description{
Returns the cell type for the given classifier.
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
cell_type = "B cells", features = selected_features_B)
cell_type(clf_b)

}
