% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_models}
\alias{default_models}
\title{Pretrained classifiers for human cells}
\format{
a list of \code{\link{scClassifR}} objects
}
\usage{
default_models
}
\description{
Pretrained classifier obtained by training Sade-Feldman dataset
and testing on Jerby-Arnon dataset.
}
\author{
Vy Nguyen, November 2020
}
\keyword{datasets}
