% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{delete_model}
\alias{delete_model}
\alias{delete_model,ANY-method}
\title{Delete model/branch from package}
\usage{
delete_model(cell_type, path.to.models = ".")

\S4method{delete_model}{ANY}(cell_type, path.to.models = ".")
}
\arguments{
\item{cell_type}{string indicating the cell type of which 
the model will be removed from package
Attention: deletion of a parent model will also delete all of child model.}

\item{path.to.models}{path to the folder containing 
the list of models in which the to-be-deleted model is.}
}
\value{
no return value, but the model is deleted from database
}
\description{
Delete model/branch from package
}
\examples{
# load small example dataset
data("tirosh_mel80_example")

# train a classifier
set.seed(123)
selected_features_T = c("CD4", "CD8A", "CD8B")
clf_t <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_T, cell_type = "t cells")

# save a classifier to system
save_new_model(new_model = clf_t)

# delete classifier from system
delete_model("t cells")
}
