% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{marker_genes}
\alias{marker_genes}
\title{marker_genes}
\usage{
marker_genes(classifier)
}
\arguments{
\item{classifier}{scAnnotatR object}
}
\value{
Applied marker genes of object
}
\description{
Returns the set of marker genes for the given classifier.
}
\examples{
data("tirosh_mel80_example")
selected_marker_genes_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
classifier_b <- train_classifier(train_obj = tirosh_mel80_example,
assay = 'RNA', slot = 'counts', marker_genes = selected_marker_genes_B, 
cell_type = "B cells", tag_slot = 'active.ident')
marker_genes(classifier_b)

}
