% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{p_thres}
\alias{p_thres}
\title{p_thres}
\usage{
p_thres(classifier)
}
\arguments{
\item{classifier}{scClassifR object}
}
\value{
Predicting probability threshold of object
}
\description{
Returns the probability threshold for the given classifier.
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_B, cell_type = "B cells")
p_thres(clf_b)

}
