% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{plant_tree}
\alias{plant_tree}
\title{Plant tree from list of models}
\usage{
plant_tree(models.file.path = c("default", "."))
}
\arguments{
\item{models.file.path}{list of models. If not provided, 
list of default pretrained models in the package will be used.}
}
\value{
tree structure and plot of tree
}
\description{
Plant tree from list of models
}
\examples{

# to create the tree of classifiers 
# (in this example, based on default classifiers)
plant_tree()

}
