% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,scAnnotatR-method}
\alias{show,scAnnotatR-method}
\title{Show object}
\usage{
\S4method{show}{scAnnotatR}(object)
}
\arguments{
\item{object}{scAnnotatR object}
}
\value{
print to console information about the object
}
\description{
Show object
}
\examples{
data("tirosh_mel80_example")
selected_marker_genes_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
classifier_b <- train_classifier(train_obj = tirosh_mel80_example,
assay = 'RNA', slot = 'counts', marker_genes = selected_marker_genes_B, 
cell_type = "B cells", tag_slot = 'active.ident')
classifier_b

}
