% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,scClassifR-method}
\alias{show,scClassifR-method}
\title{Show object}
\usage{
\S4method{show}{scClassifR}(object)
}
\arguments{
\item{object}{scClassifR object}
}
\value{
print to console information about the object
}
\description{
Show object
}
\examples{
data("tirosh_mel80_example")
selected_features_B = c("CD19", "MS4A1", "CD79A")
set.seed(123)
clf_b <- train_classifier(train_obj = tirosh_mel80_example, 
features = selected_features_B, cell_type = "B cells")
clf_b

}
