% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{simplify_prediction}
\alias{simplify_prediction}
\title{Simplify prediction}
\usage{
simplify_prediction(meta.data, mat, classifiers)
}
\arguments{
\item{meta.data}{cell meta data}

\item{mat}{expression mat}

\item{classifiers}{classifiers}
}
\value{
simplified prediction
}
\description{
Simplify prediction
}
