% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tirosh_mel80_example}
\alias{tirosh_mel80_example}
\title{A Seurat Object Sample}
\format{
a \code{\link{Seurat}} object
}
\source{
WEIZMANN INSTITUTE OF SCIENCE
}
\usage{
tirosh_mel80_example
}
\description{
A Seurat object sample was made from the dataset GSE72056,
samples corresponding to patient CY80.
}
\author{
Itay Tirosh, 2016-04-05
}
\keyword{datasets}
